API Functions and wrappers

API Functions and wrappers

api_url(url = NULL)

api_set_url(url)

api_available_models(url = api_url(), config = list(), ...)

api_get_current_beta(model_name, url = api_url(), config = list(), ...)

api_model_trace(model_name, url = api_url(), config = list(), ...)

api_model_specification(model_name, url = api_url(), config = list(), ...)

api_submit_gradient(
  model_name,
  url = api_url(),
  data,
  site_name,
  shuffle_rows = TRUE,
  verbose = TRUE,
  dry_run = FALSE,
  config = list(),
  ...
)

api_model_converged(model_name, url = api_url(), config = list(), ...)

api_setup_model(
  model_name,
  url = api_url(),
  formula = "y ~ x1 + x2",
  family = "binomial",
  link = "logit",
  all_site_names,
  config = list(),
  tolerance = 1e-09,
  ...
)

api_clear_model(model_name, url = api_url(), config = list(), ...)

api_estimate_model(
  model_name,
  url = api_url(),
  data,
  site_name,
  wait_time = 1,
  config = list(),
  verbose = TRUE,
  ...
)

Arguments

url
URL to the Plumber Server

config
additional configuration settings such as http authentication and additional headers.

...
additional arguments to send to api_submit_gradient

model_name
name of your model

data
dataset to get gradient value from. The code runs gradient_value to calculate the gradient, no individual data is submitted.

site_name
name of the site, needs to be one of the all_site_names

shuffle_rows
should the rows of the dataset be permuted, so as to decrease privacy concerns

verbose
print out diagnostic messages

dry_run
if TRUE, nothing with respect to the data is submitted to the server, but returned to see what would be submitted.

formula
model formula to fit, with tilde syntax

family
generalized linear model family, see family

link
link function to use with family

all_site_names
all the site names to fit this model

tolerance
tolerance for convergence

wait_time
Time, in seconds, to wait until to try to get new estimate

Value

The api_available_models function returns the available models running or already run.

The api_get_current_beta function returns the current beta estimates.

The api_model_trace function returns a list of the values throughout iterations of the model fitting.

The api_model_specification function returns a list of the parameters of the model specification, if the model is present.

The api_submit_gradient function returns a list from the result of the API call.

The api_model_converged function returns an indicator if the model converges or not.

The api_setup_model function submits a model to set up on the server.

The api_clear_model function clears out a model and returns the output from the API.

Examples

api_url()

#> [1] "https://rsconnect.biostat.jhsph.edu/distribglm"

api_set_url(api_url())

#> [1] "https://rsconnect.biostat.jhsph.edu/distribglm"

api_available_models()

#> named list()