Estimate the updated beta value

Estimate the updated beta value

estimate_new_beta(
  model_name,
  synced_folder,
  all_site_names = NULL,
  verbose = TRUE
)

compute_model(model_name, synced_folder, all_site_names = NULL, wait_time = 5)

model_trace(model_name, synced_folder)

Arguments

model_name
name of your model

synced_folder
synced folder to do computation

all_site_names
all the site names used to fit this model

verbose
print diagnostic messages

wait_time
Time, in seconds, to wait until to try to get new estimate

Value

A file name of the estimated values necessary for the final estimates

Examples

data = data.frame(y = c(0, 0, 1),
                  pois_y = c(4, 1, 0),
                  x2 = c(-2.19021287072066,
                         -0.344307138450805, 3.47215796952745),
                  x1 = c(-0.263859503846267,
                         -0.985160029707486, 0.227262373184513))
synced_folder = tempfile()
dir.create(synced_folder)
model_name = "logistic_example"
form_file = setup_model(model_name = model_name,
                        synced_folder = synced_folder,
                        formula =  y ~ x1 + x2, family =  binomial(),
                        tolerance = 5)
outfile = estimate_site_gradient(
  model_name = model_name, synced_folder = synced_folder,
  all_site_names = "site1",
  data = data)

#> [1] "Writing Gradient, iteration 1"
#> $gradient
#>                   [,1]
#> (Intercept) -0.3434747
#> x1           1.4794362
#> x2           0.5781000
#> 
#> $sample_size
#> [1] 3
#> 
#> $iteration_number
#> [1] 1
#> 
#> $A_mat
#>             (Intercept)         x1        x2
#> (Intercept)   0.7500000 -0.2554393 0.2344095
#> x1           -0.2554393  0.2729526 0.4265492
#> x2            0.2344095  0.4265492 4.2428652
#> 
#> $n_ok
#> [1] 3
#> 
#> $dispersion_sum
#> [1] 3
#> 
#> $u
#>                  [,1]
#> (Intercept) -0.500000
#> x1           0.738141
#> x2           3.003339
#> 
#> $deviance
#> [1] 4.158883
#> 
#> $aic
#> [1] 4.158883
#> 
#> $beta_names
#> [1] "(Intercept)" "x1"          "x2"         
#> 

estimate_new_beta(model_name, synced_folder,
all_site_names = "site1")

#> [1] "Reading in gradients, iteration 1"

#> epsiolon value

#> [1] Inf

#> [1] "/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpXftK1Y/file7ac29323904/betas/logistic_example-iteration0001.rds"

master_beta_file(model_name, synced_folder)

#> [1] "/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpXftK1Y/file7ac29323904/betas/logistic_example-iteration0002.rds"

outfile = estimate_site_gradient(
  model_name = model_name, synced_folder = synced_folder,
  all_site_names = "site1",
  data = data)

#> [1] "Writing Gradient, iteration 2"
#> $gradient
#>                   [,1]
#> (Intercept) -0.1949795
#> x1           0.8398280
#> x2           0.3281687
#> 
#> $sample_size
#> [1] 3
#> 
#> $iteration_number
#> [1] 2
#> 
#> $A_mat
#>             (Intercept)         x1         x2
#> (Intercept)  0.31498076 -0.1072779 0.09844597
#> x1          -0.10727795  0.1146331 0.17913974
#> x2           0.09844597  0.1791397 1.78189452
#> 
#> $n_ok
#> [1] 3
#> 
#> $dispersion_sum
#> [1] 0.4060058
#> 
#> $u
#>                   [,1]
#> (Intercept) -0.1192029
#> x1           0.1759771
#> x2           0.7160136
#> 
#> $deviance
#> [1] 0.7615681
#> 
#> $aic
#> [1] 0.7615681
#> 
#> $beta_names
#> [1] "(Intercept)" "x1"          "x2"         
#> 

estimate_new_beta(model_name, synced_folder,
all_site_names = "site1")

#> [1] "Reading in gradients, iteration 2"

#> epsiolon value

#> [1] 0.5638563
#> [1] "Model has converged!"

#> [1] "/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpXftK1Y/file7ac29323904/models/logistic_example.rds"

master_beta_file(model_name, synced_folder)

#> [1] "/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpXftK1Y/file7ac29323904/betas/logistic_example-iteration0002.rds"