Estimate Site Gradient

Estimate Site Gradient

estimate_site_gradient(
  model_name,
  synced_folder,
  site_name = "site1",
  data,
  all_site_names = NULL,
  shuffle_rows = TRUE,
  experimental = FALSE
)

estimate_model(
  model_name,
  synced_folder,
  site_name = "site1",
  data,
  all_site_names = NULL,
  shuffle_rows = TRUE,
  wait_time = 1,
  run_compute = FALSE,
  experimental = FALSE
)

Arguments

model_name
name of your model

synced_folder
synced folder to do computation

site_name
name of the site, needs to be one of the all_site_names

data
dataset to get gradient value from

all_site_names
all the site names used to fit this model

shuffle_rows
should the rows of the dataset be permuted, so as to decrease privacy concerns

experimental
using the glm function rather than a custom-written function

wait_time
Time, in seconds, to wait until to try to get new estimate

run_compute
if TRUE, when estimating the model, it will also try to run estimate_new_beta if all other sites have submitted. This allows all sites to be a potential computation site.

Value

A character filename of the gradient file, with the output from gradient_value

Examples

data = data.frame(y = c(0, 0, 1),
                  pois_y = c(4, 1, 0),
                  x2 = c(-2.19021287072066,
                         -0.344307138450805, 3.47215796952745),
                  x1 = c(-0.263859503846267,
                         -0.985160029707486, 0.227262373184513))
tdir = tempfile()
dir.create(tdir)
model_name = "logistic_example"
form_file = setup_model(model_name = model_name,
                        synced_folder = tdir,
                        formula =  "y ~ x1 + x2", family =  "binomial")
outfile = estimate_site_gradient(
  model_name = model_name, synced_folder = tdir,
  all_site_names = "site1",
  data = data)

#> [1] "Writing Gradient, iteration 1"
#> $gradient
#>                   [,1]
#> (Intercept) -0.3434747
#> x1           1.4794362
#> x2           0.5781000
#> 
#> $sample_size
#> [1] 3
#> 
#> $iteration_number
#> [1] 1
#> 
#> $A_mat
#>             (Intercept)         x1        x2
#> (Intercept)   0.7500000 -0.2554393 0.2344095
#> x1           -0.2554393  0.2729526 0.4265492
#> x2            0.2344095  0.4265492 4.2428652
#> 
#> $n_ok
#> [1] 3
#> 
#> $dispersion_sum
#> [1] 3
#> 
#> $u
#>                  [,1]
#> (Intercept) -0.500000
#> x1           0.738141
#> x2           3.003339
#> 
#> $deviance
#> [1] 4.158883
#> 
#> $aic
#> [1] 4.158883
#> 
#> $beta_names
#> [1] "(Intercept)" "x1"          "x2"         
#> 

clear_model(model_name, tdir)
testthat::expect_error({
outfile = estimate_site_gradient(
  model_name = model_name, synced_folder = tdir,
  all_site_names = "site1",
  data = data)
})